Move AWS E2E stack into shared harness
